requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements-raw.txt
#
-e git+https://github.com/machow/siuba.git@docs-fix-imports#egg=siuba
    # via -r requirements-raw.txt
alabaster==0.7.13
    # via sphinx
appnope==0.1.3
    # via
    #   ipykernel
    #   ipython
argon2-cffi==21.3.0
    # via notebook
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asttokens==2.0.8
    # via stack-data
attrs==22.1.0
    # via
    #   jsonschema
    #   sphobjinv
babel==2.12.1
    # via sphinx
backcall==0.2.0
    # via ipython
beautifulsoup4==4.11.1
    # via nbconvert
bleach==5.0.1
    # via nbconvert
certifi==2022.9.24
    # via
    #   requests
    #   sphobjinv
cffi==1.15.1
    # via argon2-cffi-bindings
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via panflute
colorama==0.4.6
    # via griffe
contourpy==1.0.5
    # via matplotlib
cycler==0.11.0
    # via matplotlib
debugpy==1.6.3
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
docutils==0.19
    # via sphinx
duckdb==0.5.1
    # via
    #   -r requirements-raw.txt
    #   duckdb-engine
duckdb-engine==0.6.4
    # via -r requirements-raw.txt
entrypoints==0.4
    # via jupyter-client
executing==1.1.0
    # via stack-data
fastjsonschema==2.16.2
    # via nbformat
fonttools==4.37.4
    # via matplotlib
griffe==0.25.5
    # via quartodoc
idna==3.4
    # via requests
imagesize==1.4.1
    # via sphinx
importlib-metadata==5.0.0
    # via quartodoc
ipykernel==6.16.0
    # via
    #   -r requirements-raw.txt
    #   ipywidgets
    #   jupyter
    #   jupyter-console
    #   notebook
    #   qtconsole
ipython==8.5.0
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipython-genutils==0.2.0
    # via
    #   notebook
    #   qtconsole
ipywidgets==8.0.2
    # via jupyter
jedi==0.18.1
    # via ipython
jinja2==3.1.2
    # via
    #   nbconvert
    #   notebook
    #   numpydoc
    #   sphinx
jsonschema==4.16.0
    # via
    #   nbformat
    #   sphobjinv
jupyter==1.0.0
    # via -r requirements-raw.txt
jupyter-client==7.3.5
    # via
    #   ipykernel
    #   jupyter-console
    #   nbclient
    #   notebook
    #   qtconsole
jupyter-console==6.4.4
    # via jupyter
jupyter-core==4.11.1
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
jupyterlab-pygments==0.2.2
    # via nbconvert
jupyterlab-widgets==3.0.3
    # via ipywidgets
kiwisolver==1.4.4
    # via matplotlib
markupsafe==2.1.1
    # via
    #   jinja2
    #   nbconvert
matplotlib==3.6.0
    # via
    #   mizani
    #   plotnine
    #   seaborn
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
mistune==2.0.4
    # via nbconvert
mizani==0.8.1
    # via plotnine
nbclient==0.6.8
    # via nbconvert
nbconvert==7.1.0
    # via
    #   -r requirements-raw.txt
    #   jupyter
    #   notebook
nbformat==5.6.1
    # via
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.6
    # via
    #   ipykernel
    #   jupyter-client
    #   nbclient
    #   notebook
notebook==6.4.12
    # via jupyter
numpy==1.23.3
    # via
    #   contourpy
    #   duckdb
    #   duckdb-engine
    #   matplotlib
    #   mizani
    #   pandas
    #   patsy
    #   plotnine
    #   scipy
    #   seaborn
    #   siuba
    #   statsmodels
numpydoc==1.5.0
    # via -r requirements-raw.txt
nycflights13==0.0.3
    # via -r requirements-raw.txt
packaging==21.3
    # via
    #   ipykernel
    #   matplotlib
    #   nbconvert
    #   qtpy
    #   sphinx
    #   statsmodels
palettable==3.3.0
    # via mizani
pandas==1.5.0
    # via
    #   mizani
    #   nycflights13
    #   plotnine
    #   seaborn
    #   siuba
    #   statsmodels
pandocfilters==1.5.0
    # via nbconvert
panflute==2.3.0
    # via quartodoc
parso==0.8.3
    # via jedi
patsy==0.5.2
    # via
    #   plotnine
    #   statsmodels
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.2.0
    # via matplotlib
plotnine==0.10.1
    # via -r requirements-raw.txt
plum-dispatch==1.7.4
    # via quartodoc
prometheus-client==0.14.1
    # via notebook
prompt-toolkit==3.0.31
    # via
    #   ipython
    #   jupyter-console
psutil==5.9.2
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pydantic==1.10.7
    # via quartodoc
pygments==2.13.0
    # via
    #   ipython
    #   jupyter-console
    #   nbconvert
    #   qtconsole
    #   sphinx
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
pytz==2022.4
    # via pandas
pyyaml==6.0
    # via
    #   panflute
    #   siuba
pyzmq==24.0.1
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   qtconsole
qtconsole==5.3.2
    # via jupyter
qtpy==2.2.1
    # via qtconsole
quartodoc @ git+https://github.com/machow/quartodoc.git@9a3f80ab4d0d07a553ce14eeeb24c89466c5f8ed
    # via -r requirements-raw.txt
requests==2.28.2
    # via
    #   -r requirements-raw.txt
    #   sphinx
scipy==1.9.1
    # via
    #   mizani
    #   plotnine
    #   statsmodels
seaborn==0.12.0
    # via -r requirements-raw.txt
send2trash==1.8.0
    # via notebook
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   patsy
    #   python-dateutil
snowballstemmer==2.2.0
    # via sphinx
soupsieve==2.3.2.post1
    # via beautifulsoup4
sphinx==6.1.3
    # via numpydoc
sphinxcontrib-applehelp==1.0.4
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.1
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
sphobjinv==2.3.1
    # via quartodoc
sqlalchemy==1.4.41
    # via
    #   duckdb-engine
    #   siuba
stack-data==0.5.1
    # via ipython
statsmodels==0.13.2
    # via plotnine
tabulate==0.8.10
    # via
    #   -r requirements-raw.txt
    #   quartodoc
terminado==0.16.0
    # via notebook
tinycss2==1.1.1
    # via nbconvert
tornado==6.2
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   terminado
traitlets==5.4.0
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
typing-extensions==4.5.0
    # via
    #   pydantic
    #   quartodoc
urllib3==1.26.14
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
widgetsnbextension==4.0.3
    # via ipywidgets
zipp==3.8.1
    # via importlib-metadata
